refactor DataDisplay - move graph functionality into subcomponents
