+++ src/components/DataDisplay/DataDisplay.js
@@ -1,21 +1,18 @@
-// receives the user-selected area of interest, queries the PCEX API to retrieve data
-// about indicators in that area, displays graphs of the results. Currently only displays
-// maximum temperature graphs.
-// currently the graphs are AnnualCycleGraph (timeseries API) and LongTermAverageGraph (data API).
+// receives the user-selected area of interest, queries the PCEX API to retrieve a list
+// of available, and then filters the variables by timescale (yearly, monthly, daily)
+// and sends them to display components for each group.
+// also handles user selection of model and sceanrio.
 
-
-import {testDataRequest, longTermAverageDataRequest, getMultimeta, flattenMultimeta} from '../../data-services/pcex-backend.js'
+import {testDataRequest, getMultimeta, flattenMultimeta} from '../../data-services/pcex-backend.js'
 import AnnualCycleGraph from '../AnnualCycleGraph/AnnualCycleGraph.js'
-import LongTermAverageGraph from '../LongTermAverageGraph/LongTermAverageGraph.js'
-import VariableSelector from '../selectors/VariableSelector.js';
 import React, {useState, useEffect} from 'react';
+import YearlyDataDisplay from '../YearlyDataDisplay/YearlyDataDisplay.js'
+import _ from 'lodash';
 
 function DataDisplay({region}) {
   
   const [monthlyTimeSeries, setMonthlyTimeSeries] = useState(null);
-  const [longTermTimeSeries, setLongTermTimeSeries] = useState(null);
   const [rasterMetadata, setRasterMetadata] = useState(null);
-  const [variable, setVariable] = useState(null);
   
   // fetch list of available datasets
   useEffect(() => {
@@ -28,14 +25,6 @@ function DataDisplay({region}) {
     }
   );
   
-  function selectVariable(event) {
-      setVariable(event.value);
-  }
-  
-  function dontSelectVariable(event){
-    //TODO: put something here. Ask Rod what.
-  }
-  
   // fetch data and format it as graphs.
   // currently one call to each of the 'data' and 'timeseries' APIs. 
   useEffect(() => {
@@ -46,35 +35,14 @@ function DataDisplay({region}) {
       }
   }, [region]);
   
-    useEffect(() => {
-      if(region && variable) {
-        longTermAverageDataRequest(region.geometry, variable.representative.variable_id).then(data => {
-            setLongTermTimeSeries(data);
-        });
-      }
-  }, [region, variable]);
-  
   return (
     <div className="DataDisplay">
-        <br/>
-        {rasterMetadata ? 
-          <VariableSelector 
-            metadata={rasterMetadata}
-            constraint={{}}
-            value={variable ? variable.representative : null}
-            canReplace={false}
-            onChange={selectVariable}
-            onNoChange={dontSelectVariable}
-          /> : 
-          "Loading Available Datasets"}
-        <br/>
-        {longTermTimeSeries ? 
-          <LongTermAverageGraph 
-            longTermData={longTermTimeSeries}
-            variableInfo={variable}
-            region={region}
-          /> : 
-          "Select a watershed and a variable"}
+        {<YearlyDataDisplay
+          region={region}
+          model={"CanESM2"}
+          scenario={"rcp85"}
+          rasterMetadata={_.filter(rasterMetadata, {"timescale": "yearly"})}
+        />}
     </div>
   );
 }
